tests: Remove shell pipes and redirections
Using pipes and redirections is ansible.builtin.shell allows the
occurence of pipe errors and makes the scripts more dificult to
read, as there's more 'boilerpart code' that is not nedded.

By using 'stdin' argument to this scripts the scripts bome easier
to read and to adapt to different values.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>