Use common tasks and vars for AD related tests
Some tests require AD trust and some common variables related to the
domain controller and AD users.

This patch modifies the tests that require these tasks and vars to use
the same sources. With this change configuring the variables with
inventory files is easier and ensures that data used is valid.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>