Create test scenarios with ipalab-config
ipalab-config allows the creation of complex IPA environments using
rootless containers. For example, the tool provides the means to create
an environment where a trust can be set between IPA and Samba AD DC,
which is similar enough to Windows AD DC to be used in ansible-freeipa
testing.

To start a test scenario run 'infra/scenarios/start-scenario' passing
the scenario configuration as parameter. The configuration for the
scenario will be generated with 'ipalab-config', the containers will be
started, all the nodes will be deployed and the initial configuration
will be applied to the scenario. The configuration directory will be
moved to the repository root.

Only one scenario can be run at a single time.

When the tests are finished, 'infra/scenarios/stop-scenario' will
cleanup the environment, and a new scenario can be started. The
configuration directory created will not be removed, in this case, but
will be overwritten if a new scenario is created.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>