Don't execute deployment roles if packages are not installed
When running the deployment roles, if packages are not installed, an
error of a missing module is shown, which is confusing to users.

By ensuring that either installation or uninstallation only work if a
minimal set of tools is available allows to provide a meaningful message
and let users act on the environment.

For installation, if ipa-server/client-install is not available, the
installation is aborted with a meaningful error. For uninstallation, we
assume that IPA is not installed and gently finish the process with no
errors.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>