ipahost: Fix idempotence issues with update_dns and multiple objects
When using ipahost with the 'hosts' parameters several parameter
conversion routines were not being performed causing the module to try
to execute tasks it was not suposed to.

This issues were fixed by adapting the module to use EntryFactory.

Another issue was happening when 'update_dns: true' is used with some
parameters, like 'description', 'location' or 'mac_address', where, even
if no change was to be performed the plugin would evaluate that the
arguments were different, and tried to modify the object.

The fix for this issue is to ignore 'updatedns' when comparing the
parameters provided with the existing IPA object.

This modification also fix the case where duplicate entries used with
the 'name' parameter in case 'state: absent' would fail without
indicating the use of duplicate entries.

Fixes: #1296
Fixes: https://issues.redhat.com/browse/RHEL-111063

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>